[Refactor] add directory to common component directory
