+++ src/components/common/index.ts
@@ -3,7 +3,7 @@
 export { default as BottomButton } from './BottomButton';
 export { default as HashTagButton } from './HashTagButton';
 export { default as HashTagGroup } from './HashTagGroup';
-export { default as Header } from './Header';
+export { default as Header } from './header/Header';
 export { default as P } from './P';
 export { default as Select } from './Select';
 export { default as TagButton } from './TagButton';
@@ -13,8 +13,8 @@ export { default as Icon } from './Icon';
 export { default as IconButton } from './IconButton';
 export { default as HashTag } from './HashTag';
 export { default as Layout } from './Layout';
-export { default as Modal } from './Modal';
+export { default as Modal } from './portal/Modal';
 export { default as Divider } from './Divider';
 export { default as Loading } from './Loading';
-export { default as Drawer } from './Drawer';
+export { default as Drawer } from './portal/Drawer';
 export { default as DatePickerButton } from './DatePickerButton';
